+++ LeanplumSDK/LeanplumSDK/ClassesSwift/Migration/Wrapper/CTWrapper.swift
@@ -9,7 +9,8 @@ import Foundation
 // Use @_implementationOnly to *not* expose CleverTapSDK to the Leanplum-Swift header
 @_implementationOnly import CleverTapSDK
 
-class CTWrapper {
+class CTWrapper: Wrapper {
+    // MARK: Constants
     enum Constants {
         static let StatePrefix = "state_"
         static let UTMVisitedEvent = "UTM Visited"
@@ -23,6 +24,7 @@ class CTWrapper {
         static let iOSSandboxParam = "iOSSandbox"
     }
     
+    // MARK: Initialization
     var cleverTapInstance: CleverTap?
     var instanceCallback: ((Any) -> Void)?
     
@@ -31,7 +33,6 @@ class CTWrapper {
     var accountRegion: String
     var identityManager: IdentityManager
     
-    // MARK: Initialization
     public init(accountId: String,
                 accountToken: String,
                 accountRegion: String,
@@ -48,7 +49,7 @@ class CTWrapper {
         setLogLevel(LPLogManager.logLevel())
     }
     
-    public func launch() {
+    func launch() {
         guard Thread.isMainThread else {
             DispatchQueue.main.async { [self] in
                 launch()
@@ -62,7 +63,6 @@ class CTWrapper {
         cleverTapInstance?.setLibrary("Leanplum")
         
         Log.debug("Wrapper: CleverTap instance created with accountId: \(accountId) and accountToken: \(accountToken)")
-        //CleverTap.sharedInstance()?.notifyApplicationLaunched(withOptions: nil)
         
         if !identityManager.isAnonymous {
             Log.debug("Wrapper: will call onUserLogin with identity: \(identityManager.userId) and cleverTapId: \(identityManager.cleverTapID)")
@@ -72,7 +72,7 @@ class CTWrapper {
         triggerInstanceCallback()
     }
     
-    func triggerInstanceCallback() {
+    private func triggerInstanceCallback() {
         guard let callback = instanceCallback, let instance = cleverTapInstance else {
             return
         }
@@ -85,9 +85,8 @@ class CTWrapper {
         triggerInstanceCallback()
     }
 
-    // MARK: Events
+    // MARK: Tracking
     func track(_ eventName: String?, value: Double, info: String?, params: [String: Any]) {
-        
         // message impression events come with event: nil
         guard let eventName = eventName else {
             return
@@ -163,6 +162,7 @@ class CTWrapper {
         cleverTapInstance?.recordChargedEvent(withDetails: details, andItems: items)
     }
     
+    // MARK: User
     func setUserAttributes(_ attributes: [AnyHashable: Any]) {
         // .compactMapValues { $0 } will not work on not optional type Any which can still hold nil
         let profileAttributes = attributes
@@ -182,40 +182,7 @@ class CTWrapper {
             }
     }
     
-    func isAnyNil(_ value: Any) -> Bool {
-        if case Optional<Any>.none = value {
-            return true
-        }
-        return false
-    }
-    
-    var transformAttributeValues: ((Any) -> Any) {
-        return { value in
-            if let arr = value as? Array<Any> {
-                let arrString = arr.map {
-                    String(describing: $0)
-                }
-                return ("[\(arrString.joined(separator: ","))]") as Any
-            }
-            return value
-        }
-    }
-    
-    var transformAttributeKeys: ((AnyHashable) -> AnyHashable) {
-        return { key in
-            guard let keyStr = key as? String,
-            let newKey = MigrationManager.shared.attributeMappings[keyStr]
-            else {
-                return key
-            }
-
-            return newKey
-        }
-    }
-    
-    // MARK: Identity
-    
-    public func setUserId(_ userId: String) {
+    func setUserId(_ userId: String) {
         guard userId != identityManager.userId else { return }
         
         identityManager.setUserId(userId)
@@ -231,6 +198,7 @@ class CTWrapper {
         cleverTapInstance?.onUserLogin(profile, withCleverTapID: cleverTapID)
     }
     
+    // MARK: Traffic Source
     func setTrafficSourceInfo(_ info: [AnyHashable: Any]) {
         let trafficSourceInfoMappings = [
             "publisherId": "utm_source_id",
@@ -255,6 +223,7 @@ class CTWrapper {
         cleverTapInstance?.recordEvent(Constants.UTMVisitedEvent, withProps: props)
     }
     
+    // MARK: Log Level
     func setLogLevel(_ level: LeanplumLogLevel) {
         switch level {
         case .off:
@@ -267,45 +236,4 @@ class CTWrapper {
             CleverTap.setDebugLevel(CleverTapLogLevel.info.rawValue)
         }
     }
-}
-
-public extension Dictionary {
-    /// Transforms dictionary keys without modifying values.
-    /// Deduplicates transformed keys, by choosing the first value.
-    ///
-    /// Example:
-    /// ```
-    /// ["one": 1, "two": 2, "three": 3, "": 4].mapKeys({ $0.first })
-    /// // [Optional("o"): 1, Optional("t"): 2, nil: 4]
-    /// ```
-    ///
-    /// - Parameters:
-    ///   - transform: A closure that accepts each key of the dictionary as
-    ///   its parameter and returns a transformed key of the same or of a different type.
-    /// - Returns: A dictionary containing the transformed keys and values of this dictionary.
-    func mapKeys<T>(_ transform: (Key) throws -> T) rethrows -> [T: Value] {
-        try .init(map { (try transform($0.key), $0.value) },
-                  uniquingKeysWith: { (a, b) in a })
-    }
-    
-    /// Transforms dictionary keys without modifying values.
-    /// Deduplicates transformed keys.
-    ///
-    /// Example:
-    /// ```
-    /// ["one": 1, "two": 2, "three": 3, "": 4].mapKeys({ $0.first }, uniquingKeysWith: { max($0, $1) })
-    /// // [Optional("o"): 1, Optional("t"): 3, nil: 4]
-    /// ```
-    /// Credits:  https://forums.swift.org/t/mapping-dictionary-keys/15342/4
-    ///
-    /// - Parameters:
-    ///   - transform: A closure that accepts each key of the dictionary as
-    ///   its parameter and returns a transformed key of the same or of a different type.
-    ///   - combine:A closure that is called with the values for any duplicate
-    ///   keys that are encountered. The closure returns the desired value for
-    ///   the final dictionary.
-    /// - Returns: A dictionary containing the transformed keys and values of this dictionary.
-    func mapKeys<T>(_ transform: (Key) throws -> T, uniquingKeysWith combine: (Value, Value) throws -> Value) rethrows -> [T: Value] {
-          try .init(map { (try transform($0.key), $0.value) }, uniquingKeysWith: combine)
-      }
 }
